Add a build manifest for Debian packaging
This follows a quite similar logic as for other distributions. Namely, we
build a tarball from the git repository; this is then used by Debian
scripts to build the source and binary packages using the "pkgkit"
script from https://git.sr.ht/~dlax/sr.ht-debbuilds in the "package"
task. Finally, when on a tag, the "upload" task calls the "include"
command from pkgkit which adds the built packages to the repository and
sync it back to the mirror.