Fix rendering of zero time delta
If e.g. a build completes very quickly there might might

>>> "took " + humanize.naturaltime(1).rstrip(" ago")
'took a second'

This seems fine.

>>> "took " + humanize.naturaltime(0).rstrip(" ago")
'took now'

This makes less sense.

The humanize package has a dedicated function for deltas which produces
more sensible output for this case, and is simpler:

>>> "took " + humanize.naturaldelta(1)
'took a second'

>>> "took " + humanize.naturaldelta(0)
'took a moment'

In fact, looking at the humanize source it looks like naturaltime() is
implemented as naturaldelta() + " ago" and

    if delta == _("a moment"):
        return _("now")