Fixing IntFlag/Enum value validation.
I noticed when using the API to import tickets that setting proper
status/resolution values (according to the docs) was not passing
validation. Digging into the code the issue is as follows, using
TicketStatus (class IntFlag) from todosrht.types.ticketstatus:

TicketStatus('reported')

This is invalid Python for IntFlag/Enum classes. Since the validation
checks for ValueError, it was sort of swallowing the actual issue. The
correct method to access the enum value by name is:

TicketStatus['reported']

I didn't update the ValueError check because key is already verified
above the value fetching. I think the actual try/except can be removed
in this particular case but I wasn't sure if it would affect other
corners of the platform.