Internal auth: fix invalid auth for profile req
The AbstractOAuthService.lookup_user function is used to fetch users
either from the local database or from meta.sr.ht, and the
implementation creates an Internal auth token for the requested
username. If that user does not exist, this previously tripped an
assertion failure in this code. This case now raises a 404 error, and
the assertion is made more specific.