Properly check empty/missing optional form parameters
Some request forms contain keys with empty values. For instance, logging
in from https://meta.sr.ht will generate a request containing the
"return_to" key with a value of "". Similarly, registering without
providing a PGP key generates a request containing the "pgp-key" key
with a value of "".

In such cases, calls to Validation::optional() do not behave as
expected. Since "" != None, the first two top-level if blocks are
immediately skipped in Validation::optional(), even though the second
block is where any default values are set.

In the case of logging in, this results in users being redirected to ""
instead of the specified "/".  Luckily, redirecting to "" is apparently
the same as redirecting to "/", so nothing catastrophic happens.
However, this clearly is an unintended side-effect, and not desired
behavior.