Remove allow_none parameter from Validation::optional()
The parameter isn't used by any sr.ht service and is functionally
redundant as there is Validation::required().

The if block which sets default values is also fixed. Prior to this
commit, it would only set default values for keys that did not exist in
the incoming request. For things like logging in, this was incorrect
behavior.