style: .highlight .n inherit color
Commit f66b193 introduced a contrast regression for the Pygments Name
token. This commit aims to fix that and prevent it from happening by
setting the color of those tokens to the text color of the parent
element, which is a sensible default.

`inherit` was chosen over removing the rules entirely in order to
preserve the fully-specified nature of the stylesheets from that same
commit.