fix bug where circle.svg takes up entire screen
On a cold cache, loading any sr.ht page currently results in the circle logo
covering the entire screen for around 50ms, before the page is fully loaded.

This change ensures that the logo is correctly sized from the beginning,
removing the bothersome flash and making it so that all sr.ht pages display
their content to the user around 50ms faster on first load.

Signed-off-by: Vlad-Stefan Harbuz <vlad@vladh.net>